trastero-server/requirements.txt
Flask-GraphQL
flask-pymongo>=2.0
graphene-mongo
trastero
